Adds the Vortex Luxury Emergency Shuttle (#2091)
## About The Pull Request

Adds a very expensive, yet luxury shuttle for people to buy on the
communications console. Includes a massive list of items that normally a
shuttle wouldn't have. Can be bought on the Comms Console for 300k
credits, EMAG not required.

Spare unique Head of Staff MOD's in Command's E.V.A. room.

NTC and Blueshield get CentCom MODsuits, Corporate, and Asset
Protection. (it's the emergency shuttle, the round is basically over).

Shuttle comes with fully operational atmospherics, shocked window
grilles, emergency window shutters in all rooms and garbage disposals.

Shuttle also comes with a cleanbot, hygienebot, medbot, and two secbots,
one in brig, and one in common seating.

Command have 2 Authentication Devices, 3 mini energy guns in bridge,
with 3 spare telescopic batons, a boozemat, as well as a combat medkit.
Along with their unique Tarkon protolathe reskinned to be a Command
protolathe. Not to also forget, a Request Console with announcement
access.

Medical has advanced medical kits, a SAD, advanced surgery tools,
vendors to vend supplies, a limbgrower with two spare beakers of
synthflesh, as well as a Protolathe.

Security has 2 spare batons, a set of Redsec armor, an old SecTech
(Non-Armadyne security vendor), Mindshield box, and a spare SC-2 Energy
Carbine.

Engineering has the basics, but advanced tools, a corpse and 3 SMES's
full of power to charge the window grilles. As well as 4 spare
engineering MODsuits. Engineering also has a box of materials, to insert
into the protolathes.

Atmospherics has the basics, but smart foam grenades incase of breaches
in the shuttle, RPD's for no reason, and advanced tools. As well as 2
spare Atmospheric MODsuits.

# It's been tested, all works fine.

## Why It's Good For The Game

It's a good change for the game to give crew a different means of
escape, after getting tons of money for a good job of Cargo, they
deserve a nice trip home in one of the best evacuation shuttles.

## Proof Of Testing

## After attempting to run.

![image](https://github.com/user-attachments/assets/2c257368-6508-417b-ae7e-0a0e174e0c70)

## Looking In-Game.

![image](https://github.com/user-attachments/assets/62e6b4d4-1f7e-4dbf-a2c6-3292a0230b0c)


## Changelog

:cl:
add: Added the Vortex-Class Luxury Evacuation Shuttle.
/:cl:

# Vortex-Class Luxury Evacuation Shuttle

![image](https://github.com/user-attachments/assets/fe69c642-7664-4866-9515-24cf94ee43d2)

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>
Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>
Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>