Adds *yip and *yipyip (#2254)
yip yip!

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

It adds *yip and *yipyip, because I want the kobolds to make NOISE.

## Proof Of Testing



https://github.com/user-attachments/assets/d7baa626-656e-4a46-b300-d2314fdc8796


https://github.com/user-attachments/assets/acff25e6-5061-486e-86b8-d385b92debee





## Changelog

:cl:
add: *yip
add: *yipyip
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: LT3 <83487515+lessthnthree@users.noreply.github.com>