Reverts #2263 - restoring beaker capacity to original, original values, plus changes cryo beaker for consistency (#2301)
## About The Pull Request

Reverts #2263 and changes cryo beaker to 60u

## Why It's Good For The Game

Let me be frank, the beaker nerf pr was just pushed through github with
most of the players finding out from the changelog - the reaction on
discord has been... rather negative, even the #2263 discussion and
reacion on github is mostly negative, yet it was merged anyways.

It was a wide scale nerf to both roundstart medsci and service (which
cannot print the upgraded beakers...) and it was poorly supported with
the main motivation being to make medical cook plastic... which can be
easily obtained by just asking... or in 2 different ways from cargo.
Nerfing chemistry, medicine, science and service over it is... bad and
annoying?

Reverting the changes avoids the nerfs and pointless disruption to the
game. Also more beaker space is good, as
https://github.com/Skyrat-SS13/Skyrat-tg/pull/3226 put it...

As for the cryo beaker i mostly did it for consistency with small
beakers... and maybe making someone use it.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
del: beakers are 60u/120u again
balance: changes cryo beaker to 60u from 50u
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>