Fix Tarkon Fax Machine (#2310)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Create a subtype of the fax machine that has Tarkon Command as a
contact, replace var-edited fax machine in Port Tarkon with this.

## Why It's Good For The Game

The var-edited one runtimes because `Initialize()` in fax machine code
tries to populate the CentCom contact but the variable edit has
overwritten `special_networks` with its own list (if the Tarkon one is
not intended to have CentCom listed please let me know so I can change
the PR to make that the case).

## Proof Of Testing

Trust ‼️ 
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog

:cl:
fix: The fax machine in Port Tarkon should no longer give a TGUI error.
/:cl: