Adding Jackal tail (#2311)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Taking a tail from another server and porting it over here

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Adding new tail sprites

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->


## Why It's Good For The Game
More customization, more fun
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing


![image](https://github.com/user-attachments/assets/ff4616c3-c16c-4492-a741-a3fa9ab68044)

![image](https://github.com/user-attachments/assets/a3fc1878-6301-4e49-a294-ded519811b08)

![image](https://github.com/user-attachments/assets/bae55883-9d2a-4a93-a763-e7ee7ebdfdcd)

![image](https://github.com/user-attachments/assets/a29916d2-7947-44c7-8d3b-26fbb19bd6d4)

![image](https://github.com/user-attachments/assets/461664b9-576a-4284-a929-9bf42855eabe)


<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->
<details>
<summary>Screenshots/Videos</summary>
Set above for screenshots of the new customization option. Was mainly
using this to start learning the code base and see if I could get my
development environment set up. Things are looking good on that end
</details>

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Black-Backed Jackal tail
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: LT3 <83487515+lessthnthree@users.noreply.github.com>