Removes the Security Cremator from voidraptor and replaces it with a plushie room. (#2329)
## About The Pull Request


![image](https://github.com/user-attachments/assets/45213123-000d-4e10-83a6-60703143aadd)

## Why It's Good For The Game

Plushies are objectively better and I had to find space to put them.

In all seriousness. No, security should not have their own private ling
cremator.

## Proof Of Testing

I like plushies :)

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
balance: Replaces the security cremator with a room of plushies.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: LT3 <83487515+lessthnthree@users.noreply.github.com>