Makes the queen screech not as punishing (#2371)
## About The Pull Request

This thing had the intensity of 5 and a shitton of ear damage as a
result (50). Now flashbangs deal 10 so you can imagine how bad that is.

Effectively this also made it have like 150 stacks of deafen.

closes #2261 

### Effective changes

Intensity 5 -> 3
Stun power 50 -> 80 (Effective 250 -> 240)
Damage power 10 -> 5 (Effective 50 -> 15)
Deafen power 30 -> 10 (Effective 150 -> 30) (This is still 3x of a
flashbang)

## Why It's Good For The Game

Being deaf for 5 minutes with random ringing for the next 15 is not fun
for anyone. See
https://github.com/Bubberstation/Bubberstation/issues/2261

## Proof Of Testing

N/A, this is number changes

## Changelog

:cl:
balance: Xeno queen scream no longer deafens for ages, stun times kept
same
/:cl: