Added 'Nyamagotchi' (#2335)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Adds the Nyamagotchi (tamagotchi) to the game! Take care of it! Play
with it, feed it, and let it nap... Or else it'll die! This adds the
'Nyamagotchi' item to loadout options, dorms clothesdrobes, and toy
vending machines.
This is my first time ever coding in DM! I'm pretty pleased with it.

## Why It's Good For The Game

It's just a silly little accessory, and it's interactive, too! I know
I'll at least be using and interacting with it plenty!

## Proof Of Testing

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->
Tested all outcomes and interactions.  

<details>
<summary>Screenshots/Videos</summary>


![cat1](https://github.com/user-attachments/assets/2dcd2a41-bb80-46fa-aa65-c9d10357e388)


![cat2](https://github.com/user-attachments/assets/c0c49534-4a75-4cda-bd98-dd5c1d020cf8)



</details>

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: Middiwiddi, LT3
add: Added 'nyamagotchi', a tamagotchi clone that's cute and cat-themed.
Interactive and able to attach on belt.
add: Added nyamagotchi to toy vending machine and clothesdrobe
add: Added nyamagotchi to 'pockets' loadout options.
add: nyamagotchi works as a companion for monophobia.
:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: LT3 <83487515+lessthnthree@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>