Added generic_maintenance_spawn points to Moonstation.dmm (#2455)
## About The Pull Request

Issue Summary:
On Moon Station while observing I signed up for nightmare, I was the
sole player signed up for it. once the timer ended. I did not spawn in
as nightmare, admin responded in chat:

STAFF (Sexmaster) says, "ADMIN LOG: No valid
generic_maintenance_landmark landmarks found, aborting..."

Generic_landmark_spawn locations were added to the map in areas that
satisfy the specific criteria for a nightmare to spawn.
Was able to repro the issue and fix it in the map.



## Why It's Good For The Game

It allows more antagonists roles to be played on Moonstation,
potentially seeing it played more with antag spawn points added.

Fixes #2391 

## Proof Of Testing
![Screenshot 2024-11-16
203416](https://github.com/user-attachments/assets/6a3d0ff9-3e5e-43ed-bf94-d5a6a3ca0f02)
![Screenshot 2024-11-16
203451](https://github.com/user-attachments/assets/00e2c02e-e57b-4233-842b-e2b6d7b572b6)


## Changelog
:cl: Sparex

map: Added generic_maintenance_spawn points to fix the issue with the
Nightmare not being able to spawn. They will now spawn as they should.

:cl: