Latex catsuit fixes (#2439)
## About The Pull Request

Fixes latex catsuit hiding features and also having quite possibly the
worst overlay code I've seen in this game yet.
It also isn't even handling the overlays in the proc it's meant to be
using but I only realized it now as I'm writing this. It almost gave me
a stroke so please don't ask me to put it in there I'll die

closes #1241
## Why It's Good For The Game

The code was so bad and I kinda made it so much better. Also fix bug
## Proof Of Testing

Worked on my machine (I forgot to screenshot)

## Changelog
:cl:
fix: Latex catsuit no longer hides some features
/:cl: