Revert "Dragonslayer Nerf (Pressurized Areas Only)" (#2413)
Reverts Bubberstation/Bubberstation#2388

Overcompensation and the sword got admittidely nerfed to the floor. Its
original damage values were pretty reasonable, except with the 35 armor
pen. But the issues come mostly in the minmaxing you can do with the
armor as you can still do that with any other melee item.

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>