Succumbing now alerts everyone around you that you gave up on life. (#2193)
## About The Pull Request

Per title

## Why It's Good For The Game

This will clear up some confusion in tense moments. At the end of the
day, if you succumb, it's not the other person's fault you died.
(Security shouldn't be blamed for "executing" someone who succumbs)

## Proof Of Testing

I'm not going to have an easy time screenshotting a balloon alert. If it
compiles, it works.

## Changelog

:cl:
add: Succumbing now alerts everyone that you gave up on life.
/:cl:

---------

Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>