[Greer] Replaces NTC's captain pen with a fountain pen (#2516)
## About The Pull Request
Red pen writing looks the same as the Captain's, and not as good. This
swaps the NTC's captain fountain with just a regular fountain. More
snazzy!

## Why It's Good For The Game
I get to have more style in my PDA as the NTC

![image](https://github.com/user-attachments/assets/96e596b5-f998-4801-a52e-84838216b47c)


## Proof Of Testing

<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog

:cl: ReturnToZender
del: Replaces the NTC's captain fountain pen with a regular fountain pen
/:cl:

---------

Co-authored-by: Arturlang <24881678+Arturlang@users.noreply.github.com>