Brings upstream cargo features in line with our cargo changes (#2125)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This adjusts prices and fixes a lot of oversights in my original cargo
PRs, and newly added things from upstream, helping to alleviate some of
the cargo powergame.


![image](https://github.com/user-attachments/assets/da0700bf-a483-40ad-9f12-12ce42a94d99)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
Cargo has been the premier site for weapons and armor, and while I don't
want to take away from that, some things like medical supplies take away
from other jobs, and are so cheap as to be widely available. The
paramedics want their work back.

This will restore the balance
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

![image](https://github.com/user-attachments/assets/526b097d-01be-46fd-96a4-263de8c106a9)

![image](https://github.com/user-attachments/assets/5177c8a1-3e21-4806-8980-79f8bee84763)

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: ReturnToZender
del: Removes the Lanca and Sol LMG completely from cargo orders, leaves
the Sakhno to the mosin order
del: Re-removes the ion carbine from imports (it's in the goodies tab)
del: Removes extra ammunition types for no longer used weapons from
import order
add: Removed company imports added to goodies and security orders
balance: adjusts prices on some company imports to not be essentially
free, such as medical supplies and armor
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>
Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>