Add permanent limp quirk (#2560)
## About The Pull Request
Adds a permanent limp quirk, negative with a cost of -3. Allows adding a
minor, moderate, or major limp to a characters leg that cannot be
treated short of limb replacement.
## Why It's Good For The Game
Requested by a player. Allows for players to have a character that limps
naturally without having to dislocate a leg in round.

## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>


![perm_limp_1](https://github.com/user-attachments/assets/04293b1d-7879-4875-8f28-f6421c402707)

![perm_limp_2](https://github.com/user-attachments/assets/da3f34b3-2f5b-496c-b502-be72222fd3ca)

![perm_limp_3](https://github.com/user-attachments/assets/853f5fb1-2f75-48ac-8a53-3743403dc10f)

</details>

## Changelog
:cl: Cerami
add: Adds a new negative quirk: permanent limp
/:cl:

Edit:
Changed cost to -3
Made incompatible with paraplegic and hemiplegic