Persistence Base (A new offstation syndicate map) (Should be ready for test merge) (Ready for review) (#2265)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

https://youtu.be/pPhfOGnLm08?si=QC2BPGd1pcLCXTc7 <- Updated as of 10/28

This pull request provides a replacement map for the interdyne and
dauntless offstation syndicate maps. It is supposed to be themed as a
landcrawler and features several new roles, (~~The comms operative was
removed~~) and the morale officer (The finest of waffle co. here to
enforce ~~managed democracy~~ syndicate interests. The icebox map is the
same size as the former dauntless map, while the new lavaland version is
slightly bigger to support the turrets and energy shielding.

Speaking of, one major feature of the new maps is a double layer of
energy shielding around the maps. This shielding does not recharge and
is intended to take two hits. An explosion naturally can clear both
layers at once and it is intended to prevent destruction by specific
megafauna (colossus) while also stopping miners from getting shot if
they accidentally find the ship. It does not cover the various entrances
and exits so its still worth while to be careful and if need be it
should be easy enough to attack the ship in a pinch.

**Weapon and armor wise, the ship has the following additions and
removals**

- All makarovs in the armory have been removed.
- One renostar shotgun has been removed from the armory
- Each non command crew member and the comms operative starts out with a
weapon selection beacon providing an option between a makrov, m1911,
mini egun, or energy dagger
- Each command officer barring the MAA starts with a weapon beacon
offering a choice between an ansem pistol (on par with an m1911 iirc), a
stechkin APS, or a c-20r.
- The morale officer has an additional selection of a revolver (for
executing dissidents)
- The MAA has a selection between an m-90gl with only rubber grenades
available, a bulldog shotgun, a c-20r, or a hos revolver
- An infiltrator modsuit has been added minus the psi demoralizer for
the corporate liason
- A honkerative modsuit has been added for the morale officer, it comes
preequipped with a waddle module.

**Power supply**

Per request, the ship now features a syndicate unique power source,
FISH! Make sure to pet them.

closes #2386

## Why It's Good For The Game

It has come to my attention that the interdyne offstation map which
spawns on icebox has become outdated due to the fact that it only spawns
on a single map. Rather than let this continue to be an issue I have
elected to standardize the icemoon and lavaland maps to one single map,
both to cut down further efforts of keeping up the offstation syndicate
map, and to provide a map that going forward will have room for updates.
Additionally, the new map will replace the icebox comms outpost, saving
on space for new ruins while improving (hopefully) spawning off the
offstation syndicate.

On one final note, massive mining rigs are cool and the pumpkin spice
must flow.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

See the video linked above

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: The Persistence land crawler has been deployed to local mineral
rich areas.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: LT3 <83487515+lessthnthree@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>
Co-authored-by: nikothedude <simon.prouty@gmail.com>
Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>