Expands flavor text window size and adds a scrollbar (#2631)
## About The Pull Request
This PR makes it so the TextInput component from Skyrat contains a
scroll wheel, and makes the Character Creator's flavor text preference
box bigger, so it can store more than the minimum amount of characters.
It can now show 576 characters at a time.

## Why It's Good For The Game
This makes it so the maximum size of the flavor text input is bigger
than the minimum requirement, and adds a scroll wheel for convenience!
It's now easier than ever to make a good flavor text.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

![image](https://github.com/user-attachments/assets/f70bd21e-5493-4041-a3ce-41634bb466c0)
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog

:cl: ReturnToZender
add: Flavor text in the character creator is now bigger, and has a
scroll wheel.
Fix: All Silicon preferences are now sorted together at the bottom of
the Character Lore page, for ease of access.
/:cl:

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>